[Synthesizers] Implement default overrides
Default time, fade, loop count, and sample rate may now be overridden.

Synchronized preferences strings tables. Spanish translation of new
options pending, new releases won't be pushed until they're complete.

+++ Plugins/GME/GameDecoder.h
@@ -17,7 +17,7 @@ extern gme_err_t readCallback(void* data, void* out, int count);
 @interface GameDecoder : NSObject <CogDecoder> {
 	Music_Emu* emu;
 	id<CogSource> source;
-	long sampleRate;
+	double sampleRate;
 	long length;
 	long fade;
 	NSString* codec;
